remove mocha and chai, add jest, and adjust and tidy up tests
