+++ packages/offix-cache/jest.config.js
@@ -1,0 +1,8 @@
+module.exports = {
+    "roots": [
+        "<rootDir>/test"
+    ],
+    "transform": {
+        "^.+\\.tsx?$": "ts-jest"
+    },
+}
