+++ packages/offix-cache/package.json
@@ -7,17 +7,17 @@
   "scripts": {
     "build": "tsc",
     "watch": "tsc --watch",
-    "test": "mocha"
+    "test": "jest"
   },
   "devDependencies": {
-    "@types/chai": "4.2.3",
     "@types/fetch-mock": "7.3.1",
     "@types/graphql": "14.2.3",
-    "@types/mocha": "5.2.7",
-    "chai": "4.2.0",
+    "@types/jest": "^24.0.18",
+    "@types/node": "^12.7.12",
     "graphql": "14.5.8",
     "graphql-tag": "2.10.1",
-    "mocha": "6.2.1",
+    "jest": "^24.9.0",
+    "ts-jest": "^24.1.0",
     "ts-node": "8.4.1",
     "typescript": "3.6.3"
   },
